feat(#179): split ResourceProfilePage, add AppLayout, fix client bugs
- Finding 1: Split 2008-line ResourceProfilePage into:
  - components/resource-profile/NamedResourcesPanel.tsx (self-contained panel)
  - components/resource-profile/ResourceProfileTab.tsx (profile tab JSX)
  - components/resource-profile/CommercialTab.tsx (commercial tab JSX)
  - hooks/useResourceProfile.ts (all data/mutation/handler logic)
  - utils/financialCalculations.ts (pure computeCommercialData function)
  - ResourceProfilePage.tsx reduced to 94 lines (thin orchestrator)

- Finding 2: Create shared AppLayout component with logo, nav links /
  breadcrumb, ThemeToggle, user name, sign out. All 14 authenticated
  pages updated to use AppLayout — no duplicate header JSX remains.

- Finding 3: CustomersPage + OrgsPage now use useQuery instead of
  useEffect + setState for data fetching.

- Finding 4: Fix TimelinePage useEffect missing 'timeline' dependency.

- Finding 5: Wrap localStorage JSON.parse in try/catch in useAuth.tsx.

- Finding 6: Replace data URI CSV export with Blob + createObjectURL
  in EffortReviewPage (consistent with rest of codebase).

- Finding 7: Add dropdownError state and visible warning banner when
  customer/org dropdown data fails to load in ProjectSettingsPage.
