+++ client/src/pages/DocumentsPage.tsx
@@ -4,7 +4,7 @@ import { useQuery, useQueryClient } from '@tanstack/react-query'
 import { api } from '../lib/api'
 import { getProjectCustomerName } from '../lib/projectCustomer'
 import { useAuth } from '../hooks/useAuth'
-import ThemeToggle from '../components/layout/ThemeToggle'
+import AppLayout from '../components/layout/AppLayout'
 import type { Project } from '../types/backlog'
 
 interface GeneratedDoc {
@@ -29,8 +29,8 @@ function defaultLabel(projectName?: string): string {
 export default function DocumentsPage() {
   const { id: projectId } = useParams<{ id: string }>()
   const navigate = useNavigate()
+  const { user } = useAuth()
   const queryClient = useQueryClient()
-  const { user, logout } = useAuth()
 
   const [sections, setSections] = useState({
     cover: true,
@@ -164,28 +164,16 @@ export default function DocumentsPage() {
   }
 
   return (
-    <div className="min-h-screen bg-gray-50 dark:bg-gray-900">
-      {/* Header */}
-      <header className="bg-white dark:bg-gray-800 border-b border-gray-200 dark:border-gray-700">
-        <div className="max-w-6xl mx-auto px-6 py-4 flex items-center justify-between">
-          <div className="flex items-center gap-2 text-sm text-gray-500 dark:text-gray-400">
-            <button onClick={() => navigate('/')} className="flex items-center gap-2 group">
-              <div className="w-8 h-8 bg-lab3-navy rounded-lg flex items-center justify-center"><span className="text-white text-xs font-bold">M</span></div>
-              <span className="font-semibold text-gray-900 dark:text-white group-hover:text-lab3-navy dark:group-hover:text-lab3-blue transition-colors">Monrad Estimator</span>
-            </button>
-            <span>/</span>
-            <button onClick={() => navigate(`/projects/${projectId}`)} className="hover:text-lab3-navy dark:hover:text-lab3-blue transition-colors">{project?.name ?? '…'}</button>
-            <span>/</span>
-            <span className="text-gray-700 dark:text-gray-300">Documents</span>
-          </div>
-          <div className="flex items-center gap-3">
-            <ThemeToggle />
-            <span className="text-sm text-gray-500 dark:text-gray-400">{user?.name}</span>
-            <button onClick={logout} className="text-sm text-gray-500 dark:text-gray-400 hover:text-gray-700 dark:hover:text-gray-200">Sign out</button>
-          </div>
-        </div>
-      </header>
-
+    <AppLayout
+      breadcrumb={<>
+          <span>/</span>
+          <button onClick={() => navigate(`/projects/${projectId}`)} className="hover:text-lab3-navy dark:hover:text-lab3-blue transition-colors">
+            {project?.name ?? '…'}
+          </button>
+          <span>/</span>
+          <span className="text-gray-700 dark:text-gray-300">Documents</span>
+        </>}
+    >
       <div className="max-w-5xl mx-auto px-6 py-8 space-y-6">
 
         {/* ── Generation panel ── */}
@@ -352,6 +340,6 @@ export default function DocumentsPage() {
         </div>
 
       </div>
-    </div>
+  </AppLayout>
   )
 }
