+++ client/src/pages/ProjectResourceTypesPage.tsx
@@ -2,8 +2,7 @@ import { useState } from 'react'
 import { useParams, useNavigate } from 'react-router-dom'
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query'
 import { api } from '../lib/api'
-import { useAuth } from '../hooks/useAuth'
-import ThemeToggle from '../components/layout/ThemeToggle'
+import AppLayout from '../components/layout/AppLayout'
 
 type ResourceCategory = 'ENGINEERING' | 'GOVERNANCE' | 'PROJECT_MANAGEMENT'
 
@@ -130,7 +129,6 @@ function EditRow({ initial, onSave, onCancel, saving }: EditRowProps) {
 export default function ProjectResourceTypesPage() {
   const { id } = useParams<{ id: string }>()
   const navigate = useNavigate()
-  const { user, logout } = useAuth()
   const qc = useQueryClient()
 
   const [editingId, setEditingId] = useState<string | null>(null)
@@ -218,28 +216,16 @@ export default function ProjectResourceTypesPage() {
   const availableGlobalTypes = globalTypes.filter(gt => !existingGlobalIds.has(gt.id))
 
   return (
-    <div className="min-h-screen bg-gray-50 dark:bg-gray-900">
-      {/* TODO: dark mode — add dark: variants throughout this page */}
-      <header className="bg-white dark:bg-gray-800 border-b border-gray-200 dark:border-gray-700">
-        <div className="max-w-6xl mx-auto px-6 py-4 flex items-center justify-between">
-          <div className="flex items-center gap-2 text-sm text-gray-500 dark:text-gray-400">
-            <button onClick={() => navigate('/')} className="flex items-center gap-2 group">
-              <div className="w-8 h-8 bg-lab3-navy rounded-lg flex items-center justify-center"><span className="text-white text-xs font-bold">M</span></div>
-              <span className="font-semibold text-gray-900 dark:text-white group-hover:text-lab3-navy dark:group-hover:text-lab3-blue transition-colors">Monrad Estimator</span>
-            </button>
-            <span>/</span>
-            <button onClick={() => navigate(`/projects/${id}`)} className="hover:text-lab3-navy dark:hover:text-lab3-blue transition-colors">{project?.name ?? '…'}</button>
-            <span>/</span>
-            <span className="text-gray-700 dark:text-gray-300">Resource Types</span>
-          </div>
-          <div className="flex items-center gap-3">
-            <ThemeToggle />
-            <span className="text-sm text-gray-500 dark:text-gray-400">{user?.name}</span>
-            <button onClick={logout} className="text-sm text-gray-500 dark:text-gray-400 hover:text-gray-700 dark:hover:text-gray-200">Sign out</button>
-          </div>
-        </div>
-      </header>
-
+    <AppLayout
+      breadcrumb={<>
+          <span>/</span>
+          <button onClick={() => navigate(`/projects/${id}`)} className="hover:text-lab3-navy dark:hover:text-lab3-blue transition-colors">
+            {project?.name ?? '…'}
+          </button>
+          <span>/</span>
+          <span className="text-gray-700 dark:text-gray-300">Resource Types</span>
+        </>}
+    >
       <main className="max-w-6xl mx-auto px-6 py-8">
         <div className="flex items-center justify-between mb-6">
           <div>
@@ -483,6 +469,6 @@ export default function ProjectResourceTypesPage() {
           </div>
         </div>
       )}
-    </div>
+  </AppLayout>
   )
 }
