+++ client/src/pages/TimelinePage.tsx
@@ -3,9 +3,8 @@ import { useParams, useNavigate } from 'react-router-dom'
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query'
 import { toPng } from 'html-to-image'
 import { api } from '../lib/api'
-import { useAuth } from '../hooks/useAuth'
 import { useIsDark } from '../hooks/useIsDark'
-import ThemeToggle from '../components/layout/ThemeToggle'
+import AppLayout from '../components/layout/AppLayout'
 import type { Project, ResourceType, TimelineSummary, TimelineEntry, NamedResourceEntry } from '../types/backlog'
 import GanttChart from '../components/timeline/GanttChart'
 import ResourceHistogram from '../components/timeline/ResourceHistogram'
@@ -248,7 +247,6 @@ function NamedResourcesPanel({
 export default function TimelinePage() {
   const { id: projectId } = useParams<{ id: string }>()
   const navigate = useNavigate()
-  const { user, logout } = useAuth()
   const qc = useQueryClient()
 
   const [startDateInput, setStartDateInput] = useState('')
@@ -391,7 +389,7 @@ export default function TimelinePage() {
         setEditColour(entry.timelineColour ?? null)
       }
     }
-  }, [editingFeatureId])
+  }, [editingFeatureId, timeline])
 
   const invalidate = () => qc.invalidateQueries({ queryKey: ['timeline', projectId] })
 
@@ -649,28 +647,16 @@ export default function TimelinePage() {
   const projectStartDate = timeline?.startDate ? new Date(timeline.startDate) : null
 
   return (
-    <div className="min-h-screen bg-gray-50 dark:bg-gray-900">
-      {/* TODO: dark mode — add dark: variants throughout this page */}
-      <header className="bg-white dark:bg-gray-800 border-b border-gray-200 dark:border-gray-700">
-        <div className="max-w-6xl mx-auto px-6 py-4 flex items-center justify-between">
-          <div className="flex items-center gap-2 text-sm text-gray-500 dark:text-gray-400">
-            <button onClick={() => navigate('/')} className="flex items-center gap-2 group">
-              <div className="w-8 h-8 bg-lab3-navy rounded-lg flex items-center justify-center"><span className="text-white text-xs font-bold">M</span></div>
-              <span className="font-semibold text-gray-900 dark:text-white group-hover:text-lab3-navy dark:group-hover:text-lab3-blue transition-colors">Monrad Estimator</span>
-            </button>
-            <span>/</span>
-            <button onClick={() => navigate(`/projects/${projectId}`)} className="hover:text-lab3-navy dark:hover:text-lab3-blue transition-colors">{project?.name ?? '…'}</button>
-            <span>/</span>
-            <span className="text-gray-700 dark:text-gray-300">Timeline</span>
-          </div>
-          <div className="flex items-center gap-3">
-            <ThemeToggle />
-            <span className="text-sm text-gray-500 dark:text-gray-400">{user?.name}</span>
-            <button onClick={logout} className="text-sm text-gray-500 dark:text-gray-400 hover:text-gray-700 dark:hover:text-gray-200">Sign out</button>
-          </div>
-        </div>
-      </header>
-
+    <AppLayout
+      breadcrumb={<>
+          <span>/</span>
+          <button onClick={() => navigate(`/projects/${projectId}`)} className="hover:text-lab3-navy dark:hover:text-lab3-blue transition-colors">
+            {project?.name ?? '…'}
+          </button>
+          <span>/</span>
+          <span className="text-gray-700 dark:text-gray-300">Timeline</span>
+        </>}
+    >
       <main className="max-w-7xl mx-auto px-6 py-6 space-y-6">
         <div className="flex items-center justify-between">
           <h1 className="text-xl font-semibold text-gray-900 dark:text-white">Timeline Planner</h1>
@@ -1129,6 +1115,6 @@ export default function TimelinePage() {
           )}
         </div>
       </main>
-    </div>
+  </AppLayout>
   )
 }
